requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
SQLAlchemy==1.4.31
face_recognition==1.3.0
opencv-python==4.10.0.84
Pillow==10.3.0



